Add useHandleKeyboard hook
